% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_date_to_quarter_YYYY_Q}
\alias{convert_date_to_quarter_YYYY_Q}
\title{Convert date string to "YYYY-Q" format}
\usage{
convert_date_to_quarter_YYYY_Q(date_string)
}
\arguments{
\item{date_string}{A character string representing the date (e.g., "January
2024").}
}
\value{
A character string formatted as "YYYY-Q".
}
\description{
Convert date string to "YYYY-Q" format
}
