% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{country_max_levels_df}
\alias{country_max_levels_df}
\title{Country Max Levels Data}
\format{
A data frame with columns for country names and their corresponding max levels.
}
\description{
A dataset containing the maximum levels for each country.
}
