% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{country_name_to_iso}
\alias{country_name_to_iso}
\title{Convert country names to ISO country codes}
\usage{
country_name_to_iso(country_name)
}
\arguments{
\item{country_name}{A character string representing the country name (e.g.,
"United States").}
}
\value{
A character string representing the ISO country code, or NA if the
country name is not found.
}
\description{
Convert country names to ISO country codes
}
