% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_urls}
\alias{create_urls}
\title{Create API URLs from parameter combinations}
\usage{
create_urls(processed_country_param_dt, max_level)
}
\arguments{
\item{processed_country_param_dt}{A data frame containing parameter
combinations with code columns.}

\item{max_level}{A character or integer representing the user-specified area
level.}
}
\value{
A data frame with constructed URLs.
}
\description{
This function constructs URLs for API calls based on country parameter
combinations.
}
