% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handle_default_input}
\alias{handle_default_input}
\title{Handle Default Input Values}
\usage{
handle_default_input(input, default_value, valid_values)
}
\arguments{
\item{input}{A vector of input values to be validated.}

\item{default_value}{A value that represents the default input option.}

\item{valid_values}{A vector of valid values to return if the input matches the default.}
}
\value{
A vector of input values or the predefined valid values if the input matches the default.
}
\description{
This helper function manages default input handling by returning a set of valid values
when the input matches a specified default value. If the input does not match the default,
the original input is returned.
}
