% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{indicators_df}
\alias{indicators_df}
\title{Indicators to Indicator Codes Data}
\format{
A data frame with columns for indicator names and their corresponding codes.
}
\description{
A dataset containing the names of all indicators and their respective codes
used to create URLs on the website.
}
