% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_message}
\alias{log_message}
\title{Log messages with timestamp}
\usage{
log_message(message)
}
\arguments{
\item{message}{A character string to log.}
}
\description{
This function logs messages to the console with a timestamp.
}
\details{
Helper Functions for the visiblespectrum Package

This file contains various helper functions used in the visiblespectrum package
for processing country parameters and creating API URLs.
}
